+++ src/main/java/net/pincette/json/streams/Build.java
@@ -11,14 +11,16 @@ import static net.pincette.json.streams.Common.DOT;
 import static net.pincette.json.streams.Common.SLASH;
 import static net.pincette.json.streams.Common.build;
 import static net.pincette.json.streams.Common.createApplicationContext;
+import static net.pincette.json.streams.Common.createContext;
 import static net.pincette.json.streams.Common.getApplicationCollection;
+import static net.pincette.json.streams.Common.overrideConfig;
 import static net.pincette.json.streams.Read.readTopologies;
 import static net.pincette.mongo.JsonClient.update;
 import static net.pincette.util.Util.must;
 
 import com.mongodb.reactivestreams.client.MongoCollection;
+import com.typesafe.config.Config;
 import java.io.File;
-import java.util.function.Supplier;
 import javax.json.JsonArray;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
@@ -37,8 +39,13 @@ import picocli.CommandLine.Option;
     subcommands = {HelpCommand.class},
     description = "Generates one inlined JSON per application and uploads them to MongoDB.")
 class Build implements Runnable {
-  private final Supplier<Context> contextSupplier;
+  private final Config config;
   @ArgGroup() private CollectionOrLocal collectionOrLocal;
+
+  @Option(
+      names = {"-C", "--config"},
+      description = "The configuration with overrides to the default one.")
+  private File configFile;
 
   @Option(
       names = {"-f", "--file"},
@@ -46,8 +53,8 @@ class Build implements Runnable {
       description = "A JSON or YAML file containing an array of applications.")
   private File file;
 
-  Build(final Supplier<Context> contextSupplier) {
-    this.contextSupplier = contextSupplier;
+  Build(final Config config) {
+    this.config = config;
   }
 
   private static String escapeFieldName(final String name) {
@@ -71,7 +78,7 @@ class Build implements Runnable {
 
   @SuppressWarnings("java:S106") // Not logging.
   public void run() {
-    final var context = contextSupplier.get();
+    final var context = createContext(overrideConfig(config, configFile));
     final var col =
         getApplicationCollection(
             collectionOrLocal == null ? null : collectionOrLocal.collection, context);
@@ -97,14 +104,14 @@ class Build implements Runnable {
   private static class CollectionOrLocal {
     @Option(
         names = {"-c", "--collection"},
-        description = "The MongoDB collection the topologies are uploaded to.")
+        description = "The MongoDB collection the apps are uploaded to.")
     private String collection;
 
     @Option(
         names = {"-l", "--local"},
         description =
-            "Sends the build result to stdout, even when the configuration contains a "
-                + "topology collection.")
+            "Sends the build result to stdout, even when the configuration contains an "
+                + "app collection.")
     private boolean local;
   }
 }